new features

+++ monophoton/misc/copy_local.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+TEST=$1
+
 CWD=$(pwd)
 
 cd $(dirname $(cd $(dirname $0); pwd))
@@ -9,6 +11,9 @@ LOCALDIR=$(python -c "import config; print config.localSkimDir")
 for FILE in $(rsync -nuv $SKIMDIR/*.root $LOCALDIR/ | grep .root)
 do
   echo $FILE
-  rm $LOCALDIR/$FILE 2> /dev/null
-  hdfs dfs -fs hdfs://t3serv002.mit.edu:9000 -get $(echo $SKIMDIR | sed 's|/mnt/hadoop||')/$FILE $LOCALDIR/$FILE
+  if ! [ $TEST ]
+  then
+    rm $LOCALDIR/$FILE 2> /dev/null
+    hdfs dfs -fs hdfs://t3serv002.mit.edu:9000 -get $(echo $SKIMDIR | sed 's|/mnt/hadoop||')/$FILE $LOCALDIR/$FILE
+  fi
 done
